replace casper-based js tests with selenium
