+++ tox.ini
@@ -33,15 +33,20 @@ deps =
     notebook42: notebook>=4.2,<4.3
     notebook4x: https://github.com/jupyter/notebook/archive/4.x.zip
     requests
+commands =
+    themysto install --sys-prefix
+    {posargs:nosetests -a '!uses_selenium' -v --with-coverage --cover-package=src tests}
+
+[testenv:js]
+deps =
+    coverage
+    mock
+    nose
+    requests
     selenium
 commands =
     themysto install --sys-prefix
-    {posargs:nosetests -v --with-coverage --cover-package=src tests}
-
-[testenv:js]
-commands =
-    themysto install --sys-prefix
-    {posargs:python -m themysto.jstest}
+    {posargs:nosetests -a 'uses_selenium' -v --with-coverage --cover-package=src tests}
 
 [testenv:check]
 skip_install = true
